Remove getGroupId() method from MemberListFragment
Fragments should not have public methods like this as it reduces their modularity, the current groupId is now stored in the activity.

Further changes will be required to correctly support configuration changes. That will be addressed via a separate pull request.

Set UI values in MemberListFragment in onActivityCreated().
These values require the injected members meaning that it is inappropriate to set them in onCreateView() as that is called before onActivityCreated().