Improve injection
Note: Still not happy with the SmsManager not being provided. It's only required by the release build and I don't want to make the NotifyModule, library = true. I tried having a second, release only module that provided the SmsManager to the release (non-overridden) version of the NotifyModule but the SmsManager provision was not being picked up. In the end, I just created it internally in the provideSmsTransport definition.